pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "recurrent-depth-autoresearch-harness"
version = "0.2.0"
description = "Standalone recurrent-depth autoresearch harness for continuously iterating on training, data, and evaluation ideas."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "requests>=2.32.0",
]

[project.scripts]
rdh = "rdharness.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
